+++ r7-journal-mmap/src/main/java/com/ethlo/r7/r7f/JournalAnalyzer.java
@@ -82,7 +82,7 @@ public class JournalAnalyzer implements ExchangeCompletionListener, JournalInteg
     }
 
     @Override
-    public void onChecksumMismatch(JournalExchange exchange, BodyKind kind, int journaled, int observed)
+    public void onChecksumMismatch(JournalExchange exchange, BodyKind kind, long journaled, long observed)
     {
         stats.checksumMismatches++;
         stats.problems.add("checksum mismatch " + exchange.getRequestId() + " " + kind
